fix: rename existing Address.vue to Wallet.vue
